requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudinary==1.18.2
cryptography==45.0.6
defusedxml==0.7.1
Django==5.2.5
django-cloudinary-storage==0.3.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.94.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
mock==5.2.0
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
